requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
transformers
streamlit
praw
tweepy
google-api-python-client
plotly
wordcloud
